Moved hd_bet to seperate file
